pyproject.toml
[project]
name = "scribe-extractor"
version = "0.1.0"
description = "A tool to incrementally extract text from image-based PDFs using OCR."
authors = [{ name = "Your Name", email = "your@email.com" }]
requires-python = ">=3.8"
dependencies = [
    "PyMuPDF",      # Used to read PDF files and convert pages to images
    "pytesseract",  # Python wrapper for Google's Tesseract-OCR Engine
    "Pillow",       # Image manipulation library, required by pytesseract
]

[project.scripts]
extract = "extract_text:main"